requirements.prod.txt
fastapi
uvicorn[standard]
python-dotenv
boto3
pandas
catboost
prometheus-fastapi-instrumentator
